requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-pinecone
pinecone-client
python-dotenv
unstructured
notebook
unstructured[all-docs]